--- OH-/00_Immigration/01_Immigration.py
@@ -1,26 +1,0 @@
-# 01_Immigration_.py
-#!/usr/bin/python3
-import numpy
-
-def solution(n, times):
-    answer = 0
-    smaller_value = min(times) * n
-    bigger_value = max(times) * n
-    while(True):
-        median_value = (smaller_value+bigger_value)/2
-
-        comp = 
-
-        if ():
-            answer = median_value
-            break
-        else if():
-            smaller_value = median_value
-        else if():
-            bigger_value = median_value
-
-
-    return answer
-
-if __name__ == "__main__":
-    print(solution(6, [7,10]))
